+++ PlayTogether/PlayTogether/Extensions/UILabel+Extension.swift
@@ -15,4 +15,15 @@ extension UILabel {
         attrString.addAttribute(NSAttributedString.Key.paragraphStyle, value: paragraphStyle, range: NSMakeRange(0, attrString.length))
         label.attributedText = attrString
     }
+    
+    /// 특정 글자색 변경
+    func changeFontColor(targetString: String) {
+        let fullText = text ?? ""
+        let attributedString = NSMutableAttributedString(string: fullText)
+        let range = (fullText as NSString).range(of: targetString)
+        let font = UIFont.pretendardRegular(size: 14)
+        let color = UIColor.ptGreen
+        attributedString.addAttributes([.font: font, .foregroundColor: color], range: range)
+        attributedText = attributedString
+    }
 }
